[리뷰] 호스트 계정 -> 후기관련 파일 추가 (#135)
* [init] brach 생성

* [추가] 리뷰, 답글 관련 파일 추가

* [추가] 리뷰, 답글 관련 클래스 추가. reputation클래스를 공통 클래스로 분리

* [수정] 리뷰/평점관련 리액트 페이지 진행중 : 호텔 상세에서 보이는 리뷰 컴포넌트

* [수정] 리액트 host 디렉토리 추가, 페이지 이동

* [수정] reputation뷰 생성, 관련코드 수정

* [스타일] 호텔 상세 -> 숙소후기 리스트 스타일 적용. 테스트 중

* [스타일] 부트스트랩 아이콘 의존성 추가. 코드 수정

* [스타일] 별점 표시, 리뷰내용 더보기 스타일 적용

* [수정] 리뷰 전체보기 모달 추가

* [merge] merge commit

* [수정] 리뷰의 답글 가져오기 함수 수정

* [수정] reply 관련 코드 리팩토링, 상세페이지 접기/펼치기 기능 구현 중.

* [수정] 답글 토글 스타일 적용

* [수정] console 로그 주석 처리

* [추가] 게스트 리뷰 작성(모달) 추가

* [test] merge commit test

* [수정] 리뷰 작성

* [수정] 팝업창 관련 라우트, 헤더/푸터 조건문 추가

* [수정] 게스트 리뷰 작성기능(팝업창) 완료

* [수정] 리뷰작성 팝업창 닫을 때 부모창 목록 새로고침

* [수정] 별점 자료형 수정(double ->float), 평균별점 format(소수점 2자리)

* [추가] 호스트, 후기관리 리스트 관련 모듈 추가

* [수정] review/reputation 파일명, 경로 수정

* [test] commit test

* [수정] 에러 수정

* [수정] map() 반복문 사용 오류 수정중

* [test] 반복문 에러 수정중

* [수정] 호스트 계정 -> 후기관리 관련 페이지 추가. 리뷰리스트 띄우기

* [수정] 답글 번호 매개변수 추가

+++ syFinal/src/main/java/com/example/syFinal/guest/model/dto/ReviewDTO.java
@@ -19,6 +19,6 @@ public class ReviewDTO {
 	private String rv_content;
 	private Date rv_date;
 	private int rv_re_index;
-	private double rv_star;
+	private float rv_star;
 	private int rv_hd_idx;
 }
